chore: worker + drop the windowing idea
