+++ scripts/copy-workers.sh
@@ -1,0 +1,1 @@
+mv ./dist/workers/code-worker.js ./dist/CodeViewer/worker/index.js
