+++ types/worker-loader.d.ts
@@ -1,4 +1,4 @@
-declare module 'worker-loader!*' {
+declare module 'worker-loader?inline=true!*' {
   declare const WebWorker: {
     prototype: Worker;
     new (): Worker;
